pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = "Video frame classification (PyTorch + Hydra). Training code lives in project/."
readme = false
requires-python = ">=3.10"
dependencies = [
    # Versions aligned with wheels on the PyTorch cu128 index (Linux/Windows).
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pillow>=10.0.0",
    "wandb>=0.26.0",
    "matplotlib>=3.10.8",
]

[tool.uv]
package = false

# CUDA 12.8 builds: https://download.pytorch.org/whl/cu128
# explicit = true keeps non-PyTorch deps on PyPI.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# Use cu128 wheels on Linux and Windows. PyTorch does not ship CUDA wheels for macOS;
# those platforms fall back to PyPI (CPU) automatically.
[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
